BE/model/requirement.txt
torch
clip-by-openai
Image
grpcio
ultraimport
protobuf